Fixes issue 22316: Product selector works fine after deleting its contents
The product selector of the Lines tab of the Sales Invoice window was being broken after deleting the contents of its text input. When the last character was
deleted, the SL_Invoice_Product was called with a null product, which resulted in the inpcCurrencyId to being set to ''. This property is used in the filter l
ogic of the product, and after it was set to '' the selector did not return any product.

This has been fixed in the same way that the issue 20811 was fixed, by only returning in the callout the inpcCurrencyId property if it is not null.