Fixes issue 21810: Added a property to define the client time zone
There was a problem exporting to CSV date time columns to clients who reside in areas with not fixed UTC offsets (i.e., Spain timezone is UTC+0200 in the summer and UTC+0100 in the winter). This was because just the current timezone offset was sent to the server, and when a date from another timezone offset had to
be exported, it was done incorrectly.

This has been fixed by adding a property called localTimeZoneID, which is the user has to use in a preference in order to define his timezone identifier. The
complete list of all available timezone IDs can be found here: http://www.java2s.com/Tutorial/Java/0120__Development/GettingallthetimezonesIDs.htm.

If the user does not specify a time zone id, or if he specifies an invalid one, the exporting process will keep working as it used to (dates that have an UTC offset different to the current date will not be properly exported).