Fixes issue 22481: DateTime columns can be used in a selector filter expression
OBBindings was incorrectly assuming that the format of the date filters was going to be Java's one ('dd-MM-yyyy HH:mm:ss'), but the method parseDateTime is only called from the client, and DateTime values are provided in the javascript format ('yyyy-MM-dd'T'HH:mm:ss').

Apart from that, the dates, that arrive as an UTC date, are converted to the server timezone.