Fixes Issue 0022625: Issue in auto save in grid view
When a process that is executed for an unsaved record in grid view tries to access the record an error is generated because the process is executed before saving the record. Modified code to save and then execute the process.

--HG--
extra : rebase_source : de97c6214871f38f229549a2a2e5b6282ea8e682