Fixes issue 22690: Save button is enabled in Return to Vendor window
The save button was not being enabled because there was one mandatory field (grandTotalAmount) that was empty. The thing is that this field belongs to the status bar, so it is read only, and should not be taken into account when cheking if all the mandatory fields have been set.

The allRequiredFieldsSet function has been modified so that it doesn't take into account the fields that are not displayed in the form (i.e. the status bar fields)