Fixes issue 22712: Status bar fields with display logic are not duplicated
Status bar fields withe display logic were being shown in the form because smartclient uses the showIf content to check if the field has to be printed in the form. Status bar fields do not have to be printed in the form, so they can't use showIf, instead they are going to use a new attribute called statusBarShowIf.