Fixes issue 5202: Implemented display logic for tabs
Now the tabs can define display logic, based on the values of its ancestor tabs fields.

Summary of the changes:
-DynamicExpressionParser.java: Adds a new attribute called inpColumnNames, that is only used in the tabs display logic and that will be used to convert the tok
ens to its inp column format. This needs to be done because the display logic is going to feed from the view context, and there the ancestor tab fields are sto
red using its inp column name.

-ob-view-tab.js.ftl: Adds to the view description of the tab the showTabIf property, but only if the tab has a display logic defined.

-OBViewTab.java: Adds the getShowIf method, that is used in ob-view-tab.js.ftl to retrieve the display logic of the tabs

-ob-view-form.js: Calls the updateSubTabVisibility (defined in ob-standard-view.js) whenever a fields changes and loses focus.

-ob-standard-view.js: In the first chunk it wraps the tab display logic in a try block to prevent errors if the display logic is not defined properly (this same thing is done for the fields display logic). The second chunk defines and uses the updateSubTabVisibility function, that based on a view will be used to determine the visibility of all its subtabs. The function is called whenever a new record is selected.
If a subtab that was active in its tabBar is hidden, the first visible tab of that tabBar will be selected, if any. If there are not any remaining visible subt
abs, the whole tabBar will be made invisible.

-ob-standard-window.js: Changes have been made in the functions to move among tabs to take into account that it should not be possible to move to a hidden tab.

--HG--
extra : rebase_source : 94d20be2c397994e0429f0f376ced2c5faef9e1a