Fixes issue 22753: Tab display logic works with preferences
Preferences were being treated as fields, so its name was wrongly being converted to its inp form. Of the three types of variables that can be used in the display logic of a tab (session attributes, preferences and fields of the tab ancestors) only the fields have to be converted to its inp format. Both session attibutes and preference values can be included in the tab definition, because they are independent of the selected row.