Fixes issue 22738: Enables applying the same dataset to different organizations
The getOrgIds nethod of the EntityResolver class has been backed out to the version before the changeset that causes the regression.

Now when the EntityResolver is used from the DAL webservices, it will not try to look for the translated IDs, it will just try to get the item with DAL,using DALs client and organization checks.

--HG--
extra : transplant_source : %85%257%2A%1F%CEpG%EE%3B%AE%AD%23%E1%DD%FC%20%8Bm%D9