Fixes issue 22727: Status bar fields are shown even with a default saved view
Before the fix, the hiddeninform property was used in the statusbar fields personalization to decide if they should be shown. But this property was set to true for all statusbar fields to prevent them from showing in the form, so this check has been removed.

--HG--
extra : transplant_source : %90xL%1F%26%00%BF%B0%2C%9D%01%E0%D2%3A_%92%7FbOr