[pw] Changed the string to indicate that a process can be executed again.
If there is a controlled exception in the backend and the process is allowed to be executed again (i.e. validation error), the backend must add the following line to the returned message: message.put("retryExecution", "Y").

Previously instead of "retryExecution", "validationError" was being used. It has been changed because "retryExecution" is more generic.