Fixes issue 22801: ID of product not shown if the full product name is entered
The problem was that in the changed function the setElementValue was being called passing newValue as an argument. If the product name is not entered completely, then newValue will be the incomplete name and nothing wrong will happen. The problem is that if the full name is entered, newValue is the product ID, and when it is passed as an argument to setElementValue it is shown as is.

This has been fixed by passing the newValue to the mapValueToDisplay function. If newValue is a partial name the function will return the partial name, and if newValue is the product ID, it will return the product full name.

--HG--
extra : transplant_source : %0CK%A1P%0F8AWF%E9%92%7D2%C7i%07%1B%D4%B3%3D