Related to issue 22801: Refactors fix to prevent regression
The fix has been refactored, because the previous approach was causing a regression. Now, the mapValueToDisplay will only be used if the newValue is an UUID, and that function will not be used in the filterDataBoundPickList function (it was not needed).

--HG--
extra : transplant_source : %86s%83%D4%DA%B0%D1%D6b%EC8%86%E6%D7%E3%1E%F4%C5%B3%06