Fixes issue 22885: Dates in filters are serialized as Date, and not DateTime
The problem was that the Date was being serialized as a DateTime, and javascript was converting the resulting datetime to the UTC format.

I.e. if the user entered 24-01-2013, the date sent to the server for filtering was 23-01-2013 23:00:00. This conversion was being done when smartclient serialized the date. If the date had the logicalDate property set to true, then the conversion was done properly.

The change has been done only in the ob-formitem-minidaterange.js to ensure that the DateTime fields are serialized with both its date and time parts.