Fixes issue 22521: In P&E the selection is not removed when sorting the grid
When a record was selected by clicking on an area other than the checkbox, the record was deselected when a column was sorted. This was caused because the selectionUpdated function, that was in charge of keeping the list of selected records, was only being called when the record was clicked on its checkbox.

This has been fixed by executing that function inside the selectionChanged function, that is executed when a record is selected, regardless if the click has been done in the checkbox.