fixed bug 22645: tabs without explicit reference to parent id are not refreshed
The problem was as the subtab was not explicitly setting its parent record (it
was done through where clause), when changing selection in parent an empty
criteria was created preventing request to datasource of being done.

The fix creates for this case a dummy criteria to force request to be triggered.

--HG--
extra : rebase_source : d1bccf2131c4cfd43afdc181d8353027e629140b