Fixes issue 22522: In P&E unselected records remain unselected after sorting
Before this fix, if a record was originally selected when a P&E window was opened, if it was deselected and a column sorted, the record was marked as selected again. This has been achieved by keeping a list that contains all the deselected records. When the data is reloaded, it a record is present in this list, it will not be marked as selected.

--HG--
extra : rebase_source : 27206cce2266b3075e09952ed9f52b71bb604468