Fixes issue 22750: Number format works with decSeparator = ','
Using the format file attached to the issue https://issues.openbravo.com/view.php?id=22750, when a number higher than 999 was entered in a numeric filter, only the last 3 digist remained (i.e., if 1234 was entered, only 234 remained).

This has been fixed by including the group separator and the default  grouping size in the call to JSToOBMasked used to format a numeric value.