Fixes issue 22810: Record is not opened in form view, message is not lost
The problem was the after opening a tab using the openDirectTabView function, the first time the ob-view-grid.delayedHandleTargetRecord function was called, the selected record would open in the form view and the grid message would be lost. This behaviour is controlled using the isOpenDirectModeLeaf property.

This property only makes sense when the openDirectTabView function has been called with a targetRecordId (a specific record will be opened). To fix this issue, when the data arrives to the grid, the isOpenDirectModeLeaf property will be deleted unless the property targetRecordId has been set.