Fixes issue 23001: Clearing the filter works properly in P&E windows
The problem was that if in a P&E window the user entered something in a filter and then deleted that filter, the application considered that there were no filter active, so clearing the filter did not even try to call the datasource. The problem is that even though there were no filter, that window had a filter clause that needs to be cleared if the user clicks on the Clear Filter button.

This has been fixed by forcing a reload when the user clicks on the Clear Filter button by adding a dummy criterion to the criteria.

--HG--
extra : rebase_source : afb96981e13a8d17683da0a0f1f2ec9cb585baf6