Fixes issue 22971: Columns with null values can be exported to CSV
The problem was that the convertPrimitiveValue method of the DataToJsonConverter class was returning null when a null object was provided. If then that null object is used in the JSONObject.put(key, value) in the value field, the field was not added at all.

This has been fixed by returning an empty string instead of a null object.