Fixes issue 22821: Triggers executed after entering full identifier in selector
The problem was that if the full identifier of the selector option was entered and then another field was clicked, the pickValue() method of the selector was not called and the triggers were not executed. This has been fixed by detecting the case where the whole identifier has been entered and if the user blurs out of that field, the setValueFromRecord is executed and the callouts were executed accordingly.

--HG--
extra : rebase_source : 39cd49e0885111f62a15dc1d3fa36d39f0d2eb34