Fixes issue 22903: Focus is not lost when changing images in Client window
When the Upload button of the popup was clicked the processEvent function of ob-eventthandler was executed. This function, among other things, tries to set as active view the view that owns the button. The problem was that the view being set as active was not the pop up, but the header tab of the window. This has been fixed by preventing this from happening where the window where the event originates is a popup.