Fixes issue 22766: Information is not lost when saving a record with errors
The problem was that when a record was saved and in the saving process it was returned with validation errors, the record was being cleared from the grid. This happened because the List
Grid recordHasChanges function returned false (the record had changes indeed, in fact it was not yet saved in the database).

This has been fixed by overwritting the recordHasChanges function, so that it always returns true if the record had validation errors. In that case we can be sure that the record has ch
anges, because we know that a record with validation changes can not be saved in the database.

--HG--
extra : rebase_source : f9f09e645f0d91da171afa4f62611900924cd5ab