Fixes issue 22767: Info is not lost when updating personalization settings
When Windows Personalization is created, the 'Visible at' fields (organization, role, etc) are sent to the window personalizationa handler, where the personalization is created properly. The problem is that when an existing personalization is updated, the 'Visible at' fields were not being sent as parameters, and the window personalization handler was setting them to null.

This has been fixed by only setting the 'Visible at' fields when the record is being created. When it is updated, the Visible at info should not be changed (see [1]), only the personalization data should be modified.

[1] https://code.openbravo.com/erp/devel/pi/file/f721f5a384cb/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/personalization/ob-personalize-form.js#l594

--HG--
extra : rebase_source : d7a0e78b43c45d5d0f105c69eecc6db3313cbad9