Fixes issue 23068: Sorting is not reset after saving a record with errors
This problem was caused because the value of the computed column changed from null to an empty string. This caused the sorting of the grid to be cleared, due to [1]. The change from null to an empty string is done in the fixNull250 function. This functions is needed for evaluating properly the display logic of the field, but it should not be called using the record itself, but with a clone of the record, so that the record itself is not changed.


[1] https://code.openbravo.com/erp/mods/org.openbravo.userinterface.smartclient.dev/file/8c2bc26694da/web/org.openbravo.userinterface.smartclient/isomorphic/client/widgets/ListGrid.js#l28665