Fixes issue 23049: Makes pin column of multi selectors not filterable
The undesired checkbox appeared in the second column of the filter because the pin column was actually filterable, when it should be not filterable. Now that the column is not filterable the checkbox no longer appears in the filter.

--HG--
extra : rebase_source : 28751c26f4d97b9a1cd006fb0b9bc9c70df8feb2