Fixes issue 22821: Triggers executed after entering full identifier in selector
This changeset is the same as [1] with one exception, we don't want to set the fullIdentifierEntered property to true when the user selects an option using th
e selector window, because in that case setValueFromRecord has already been executed, so there is no need to do it again in the blur function. This has been f
ixed by not setting the fullIdentifierEntered property if the change event was originated from the setValueFromRecord function.

[1] https://code.openbravo.com/erp/devel/pi/rev/98421b2454605ceb4c47e1267a1787c74d82e169