Fixes issue 22767: Info is not lost when updating personalization settings
Fix has been refactored and improved:
- When a personalization is going to be stored, the parameters carry a new parameter called applyLevelInformation. If the value of this parameter is false, th
en the level information (client, organization, role and user) will not be fetched from the request, but from the existing personalization. The only point in
which this parameter is false is when a Form personalization is updated.

The changes made in the PersonalizationHandler class have been reverted, they are not needed anymore.

--HG--
extra : rebase_source : c4d33eda2309f164462887207ba0fcb06e4508fb