Fixes issue 23154: Read only logic is working as expected.
In the fix of issue 23068 a change was made so that the current values of the form were not changed to '' if they were originally null. The problem is that the read only logic depended on this to work. This has been fixed by fixing the values inside the onFieldChanged function, the same way it was fixed for the display logc in the issue 23068.