Fixes issue 22638: Enables using display fields in selectors based on views
The root problem was in the implementation of DalUtil.getValueFromPath. Two parameters are passed to that function, a BOB and a propertyPath. In selectors based on tables, the property always belongs to the BOB, but this is not always true for selectors based on views. I.e., in the Product selector, if the Product field is selected as value field and the Name field as display field, the following call will be made:

DalUtil.getValueFromPath(productBOB, "product.name").

The product BOB does not have a product (first part of product.name) property, but the view has. This has been solved by making recursive calls to getValueFromPath, stripping each time the first part of the propertyPath. This would be the execution for the previous example:
DalUtil.getValueFromPath(productBOB, "product.name")
   - The productBOB does not have a product property
   - return DalUtil.getValueFromPath(productBOB, "name")
      * The productBOB does have a name property, it returns its value