Fixes issue 22523: It is possible to use a selector reference based on a view
The problem was that the referenced property of the column that used a selector reference based on a view was wrong. I.e. if a selector is created using the view m_product_price_warehouse_v, if the value field used is its product.id property, it would not be possible to save an order line using this selector because DAL would not be able to retrieve the product, it wo
uld try to fetch a record from the m_product_price_warehouse_v view using the product id, which is not the same as the view primary key.

To fix this, the user must set the column field of the defined selector so that it points to the entity used in the value field. A change has been done in SelectorDomainType.getForeignKeyColumn to ensure that this works with views.