Fixes issue 23045: Date parameters do not generate tomcat error in PW windows
The way the width of the parameters is fetched has been refactored. Now it depends on the UIDefinition of the parameter. The default width is '*', and if a subclass of UIDefinition needs to return a different width, it must overwrite the new public String getParameterWidth(Parameter parameter) method. I.e., DateUIDefinition overwrites it like this:

  @Override
  public String getParameterWidth(Parameter parameter) {
    return "50%";
  }

--HG--
extra : rebase_source : e752d3152a6b5751603a1bf9b9df6d2757128a16