Fixes issue 23203: Filter works with DateTime fields
With the fix of https://issues.openbravo.com/view.php?id=22885, when a Date or DateTime column is filtered, plain Dates are send as criteria. This is done to prevent Date from being converted to its UTC format. The AdvancedQueryBuilder was not prepared for this change, and was expecting a full Date and Time value for DateTime columns. Now it first try to parse the value as a DateTime and if it's not possible it tries to parse the value as a Date.

--HG--
extra : rebase_source : 31a7139096d3a58402b441784d42127311fd1bc7