Fixes bug 22655, fixes bug 22851. Display logic works correct with one record.
When a record is clicked, it is selected and a call to the FIC is done in SETSESSION mode. This call returns with the form session variables. The problem is th
at when there is only one row in the grid, it is automatically selected and no call to the FIC is done in SETSESSION mode.

This was happening in this issue, so the button whose display logic used session variables was not working properly. Only when the grid was refreshed the call
to the FIC is SETSESSION mode was done.

This has been fixed by making a call to updateButtonState(false, false, true) when data is received from the dataset being returned only one row. This call wil
l result in a call to the fic in SETSESSION mode. The first argument ensures that the SETSESSION call is made, the second one is not used and the third one ens
ures that the call is made even if the view is being updated is not the active one.

--HG--
extra : rebase_source : 9eefedc1d1146ccbdd34c0fb0c4993c119c583b4