Fixes issue 23220: Display logic works as expected
In the fix of issue 23068 a change was done to prevent all properties of the form being changed from null to ''. This change had to be done on a clone of the record, not on the record itself.

The problem is that at one point the record itself was being used, being taken for granted that its null values would be replaced with an empty string. This has been fixed by replacing the record by the clone that already has had its null values replaced by ''.