Fixes Issue 0023155: Special characters not recognized when filtering in grid view
If the character to be compared is a special character, an escape character '|' was added.
But if the character is '|' , and if the type is char, appending a char '|' again results in adding its ascii value (124) so the text is modified as "filter_string"248.
Modified the type of escape character to string for the this replace alone. The rest of the code uses the char definition of escape character.

--HG--
extra : rebase_source : 96a7b1b70747a02165192e1b99d42a37e9ab299e