Fixes issue 23258: Enables context attributes as default values in PW windows
Now it is possible to use a context attribute as default value for a parameter in PW windows. For example, if a PW window is opened from the Sales Order window, @C_BPARTNER_ID@ can be used as a default value for a parameter, that will be initialized to the business partner of the record that was selected when the PW window was opened.

--HG--
extra : rebase_source : 5b48e188a741c34bf82f6c373186a9a5149d6891