Fixes issue 23260: Time fields are exported properly to CSV
In the fix of the issue 21810 [1] a change was done to export the DateTime columns properly to CSV. This fix did not apply to Time fields, so when a time column was exported to CSV it could belong to the wrong timezone.

This fix applies the fix of the DateTime columns to the Time columns.

[1] https://issues.openbravo.com/view.php?id=21810

--HG--
extra : rebase_source : 583f2183c7b856bea7564b0926ae34ed497f4483