Fixes issue 23006: Fixes simultaneous calls to the datasource when filtering
Before this fix, if a call to the datasource was done before the response of the previous call was received, the response of the second call would not be applied. The second response was not applied because of this code [1]. This code handled a special case that happened when the user scrolled the grid to see more records. There is no need to execute this code when the origin of the call the the datasource is a filter request.

This has been fixed by including a parameter called 'filtering' in the datasource request only when its origin in the filter. Then, the code [1] is only executed when this parameter is not set to true.

[1] https://code.openbravo.com/erp/devel/pi/file/tip/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/grid/ob-view-grid.js#l248

--HG--
extra : rebase_source : e690df949cea7efa0f628751ecc85963068b7870