Fixes Issue 0023267: Error using fields with selectors as identifier
When an external module field that uses a selector is set as identifier, ComboTableData errors out.
The comparison was done with the external module field 'EM_DBPREFIX_FIELDNAME' in the base table where the field name was FIELDNAME.
Modified code to fetch the column name from Selector definition in case of an external module field.