Fixes issue 23334: Product selector fetched the proper value of @ad_org_id@
The problem was that the product selector did not send the windowId to the datasource. If there is no windowId, the datasource does not try to fetch the session variable from the window, as it should. In this particular case the selector should try to fetch the ad_org_id associated to the window, but instead is fetching the ad_org_id of the organization the user is logged with.

Selectors are not only used in generated windows but also in parameter windows. In that case it will not be possible to send to the datasource any windowId, but the process should not fail either.

--HG--
extra : rebase_source : edfe5b9fa7a1fabc6f8e9bc77b96331ffd67dfa4