Fixes issue 23082: Removes smart sprites functionality
The root cause of the fix is that there two versions of the google collections library. The old one, located in lib/build/google-collections.jar was only used for the smart sprites functionality, that only applies to 2.50 windows and to the login window. This functionality does not work with the neweste version of the google collections library (modules/org.openbravo.base.weld/lib/runtime/google-collect-1.0.jar).

It has been decided that this feature is going to be removed. It has almost no significance in Openbravo 3.0, in fact it is disabled by default. Changes done:
- Removed now unused libraries:
   * lib/build/commons-math-1.2.jar
   * lib/build/google-collections.jar
   * lib/build/smartsprites-0.2.1-alpha.jar

- Removed functionality from src/build.xml

- Removed configuration option from Openbravo.properties.template

- Updated legal folder