Fixes Issue 23124, Related to Issue 23267 : Selectors as Identifiers
When comparing selector columns in ComboData the column name in the table was used to compare with the selector table. This results in exception when the column names are different.
In case the selector has the column field, that column is used for comparison avoiding this error.

--HG--
extra : rebase_source : a26ff3e469ba0d06e72d645bf80610a3445da083