Fixes issue 23418: It is possible to include core messages in the I18N object
A new flag has been added to the Message header tab, that will only be shown if the selected message belongs to the core module. Checking this flag will result in the inclusion of the message in the OB.I18N object, which can be used to retrieve the label from the client side. By default labels that belong to a module other than core are included, so this flag is not needed in those cases.

--HG--
extra : rebase_source : 37cb8823c23ea3a3abfbbca29c093a426dc6ad22