Fixes issue 23479: Sending email works having several email configurations
It was not possible to send an email having defined several email configurations. It did not work if a client had two configurations defined, or even it two different clients had one configuration defined.

To fix this a client filter has been set (so the email configuration of another client is not fetched), and the results are not fetched using uniqueResult, just in case a client has defined several email configurations for the same organization.

--HG--
extra : rebase_source : 1059483aee060ed92e8710cd3190b9b5b9dd1586