Fixes issue 23294: Parameter windows show messages with title and text
The function handleResponse was not showing the messages properly. At that point, the severity and the message text are available, and the message was being displayed like this:
messageBar.setMessage(message.severity, message.text). The first argument of the function is the severity, the second the message title and the third one (not used here), the message text. The way it was being called, the message text was shown in the message bar title.

To fix this problem, the function is now used like this:
messageBar.setMessage(message.severity, null, message.text); This way, the message title is retrieved using the severity, and the message text is shown in its place.