Fixes bug 22997: Manual processes are not always disabled using SecuredProcess
When the SecuredProcess preference was set to 'Y', all manual processes were being disabled, and only the defined processes (new ones) that the user has explicit access to were enabled. This has been fixed, so the manual processes that the user has been given access to will also be enabled.