Fixes issue 23011: Enables using simple functions in HQL order by clause
Now it is possible to use one argument functions (i.e. abs(), max(), min(), etc) as part of the HQL order by clauses. A regular expression is used to check if the orderByParam matches the pattern functionName + '(' + property + ')' + 'desc' (optional). If so, it retrieves the function name and the property name, and resolves the property name. After doing it, it wrappes the resolved property name in its function and adds ' desc' if the function was meant to be sorted in descending order.

--HG--
extra : rebase_source : 62bb131fa193be74492187af1dca53bdfb12bbdb