Fixes issue 23492: Fixed definition of the OBUIAPP_Process_Access.Role column
The Role column of the OBUIAPP_Process_Access table was not defined as IsParent. That meant that when the context info is sent to the datasource, the value of the AD_ROLE_ID property is the ID of the role the user has logged in, instead of the id of the role selected in the Role tab.