Fixes bug 23474: Selectors work with identifiers that contains carriage return
Before this fix, if the identifier of a record provided by a selector contained a carriage return, the selector was not able to obtain the record id using its identifier. This happened
because the identifier was stored in the valuemap with its carriage returns, but then the identifier passed to retrieve the record ID did not contain the carriage return, so there was n
o match.

This has been fixed by removing the carriage returns before storing the identifiers in the value map.

--HG--
extra : rebase_source : db9aa3c53873b778ead78449bae45421d6ed6b49