Fixes issue 23431: It is possible to send several emails at the same time
The pocData attribute of the PrintController class stores the email attachments. The problem was that this property was global and shared among all threads, so it was not possible to send several emails at the same time, because in that case the attachment sent in the concurrent emails was the one generated the last time the Send Email button was pushed.

This has been fixed by storing the pocData of each email as a session attribute, using as a key the document number and the table name. This way pocData is no longer shared among threads, each thread has its own instance of pocData stored as a session attribute. Once the email is sent, the pocData is removed from the session.