Fixes issue 22246, fixes issue 22073: Prevents false validation error
When a new line is created while another line is being edited, the function hideInlineEditor of ob-view-grid is called. In this method the values of the form are cleared. The root cause of this problem is that even though the values of the fields are cleared, the property _editorEnterValue of selectors and fk items is not cleared. This results that the value of those fields is different than the value of its _editorEnterValue property, so the validation is done, and in this particular case it fails because the field is mandatory and its value is null.

This has been fixed by clearing the _editorEnterValue property of selectors and fk fields when their values are being cleared.

--HG--
extra : rebase_source : f183fa1455fcac61ea9b5c0e8b798d544d5c2cdf