Fixes issue 23433: Modified date autocompletion to support MM-DD-YYYY format
The date autocompletion has been modified. Before this fix, If only one or two digits were entered, they represented the first date part of the format (number of days in DD-MM-YYYY format, number of months in MM-DD-YYYY format). Now, if only one or two digits are entered they will represent the number of days, regardless of the date format being used. The rest of the date parts will be autocompleted as usual, using the current month and year.