[DS Based Tables] Added flag to tell a datasource is used as table data origin
A new boolean column has been added to the OBSERDS_DATASOURCE table. This flag indicates whether this datasource is used as a data origin in the definition of a Table. Only datasources with this flag checked will be shown in the new Datasource selector used in the Datasource column of the Table table.

If a datasource has this new flaw checked, the subtab Datasource Fields will not be shown. This tab is hidden to avoid confusion, because the columns returned by the datasource that is used as a data origin are defined in the Columns tab of the Tables and Columns window.