Related to issue 20422: Updates the ad_create_audit_triggers SQL function
The ad_create_audit_triggers SQL function has been updated to use the new AD_TABLE.ISAUDITINSERTS column,

The behaviour of the triggers when a record is updated or deleted has not been changed. When a record is inserted, if the ISAUDITINSERTS column of the audit column is'Y' (default value), the behaviour will be the same as the previous one. If ISAUDITINSERTS is 'N', then only the key column of the audited table will be inserted in the audit table.

--HG--
extra : rebase_source : 80b36a7a6453fcc480896144827939d8b9a571a2