Fixes issue 20422: Updates the AuditTrailPopup to consider show ID column
The AuditTrailPopup has been updated to show the ID column of the audit table. This is needed because now if an audited table is flagged as ISAUDITINSERTS='N', only the ID column of the table will be inserted in the audit table. Originally the popup did not show the records of the audit table belonging to its id column.

--HG--
extra : rebase_source : 21e75d7d3c513e4bb2483ed7a162ca92a1e2828a