[DS Based Tables] OBViewTab returns the proper dataSourceId
The getDataSourceId method of OBViewTab used to return always the name of the table. This has been changed, and not if the table is based on a datasource it will return its associated datasource id.

A new method, public boolean isDataSourceTable(), has been defined to be used in ob-view-tab.js.ftl.