[DS Based Tables] Adds getPropertyByTableId method to the DalUtil class
Datasource based tables does not necessaraly have a database table name, because it does not need a table in the database. DalUtil provides a method to retrieve a property based on a table name and a column name, so it could not be used for datasource based tables. A new method has been created that uses the table id instead of the table name to retrieve the property.