Fixes issue 23675: OBContext initializes properly the user default warehouse
If the user did not have a default warehouse defined, OBContext was not initializing the warehouse properly. Instead of using the previosly obtained default organization to fetch the warehouse, it was using any organization in the user accessible tree. This way, the default warehouse could belong to a organization differente than the default organization.

To fix this, if a default organization is available, it will use it. If there is no default organization available or if there a no warehouses that belong to the default organization, the user accessible organization tree will be used.

--HG--
extra : rebase_source : 13859016d606cfd0ef238dd1c2f350a1500674d3