Related to issue 23761: Fixed destroyAndRemoveMembers function
There was a bug in the destroyAndRemoveMembers. This function is used to destroy and remove a list of members from a layout. When this function was used ~50%
of the members of the layout where not being destroyed, causing leaks.

--HG--
extra : rebase_source : 4606a79d9a32b599bff2e99e11588487fae12332