Fixes issue 23801:Inserting several entities in one web service call works
It was not possible to insert in a same web service call an entity and one child. To support this, the classes DalWebService and EntityResolver have been modified. The approach using the lookForTranslatedIDs [1] has been reverted. Now, the EntityResolver.resolve method first tries to resolve the entity as it used to do it before the regression was introduced, and if it does not find the entity then it tries to fetch it using an OBCriteria without applying the organization filter, so that this issue [2] is not reproduced.

A new function has been created to search for the entity. It accepts a boolean that will be used to specify if the organization filter should be applied. The old doSearch method calls the new method passing true to this flag, so it keeps working as usual.

[1] https://code.openbravo.com/erp/devel/pi/rev/8525372a1fce7047ee3baead23e1ddfc208b6dd9
[2] https://issues.openbravo.com/view.php?id=21716