Fixes issue 23758: Amount and price columns are properly rounded
When a numeric field loses the focus its value is going to be rounded using the number of decimal digits specified in its type instance. Then in the saveEditedValues of ob-view-grid, th
e new values are going to be recomputed to make sure the updated values are being retrieved.

--HG--
extra : rebase_source : 6dc997f5307bbeeb6c038faa101931867e6b3000