Related to issue 23078: ID of the default module is no longer stored in session
When a module is set as default, its ID used to be stored in session with the @AD_MODULE_ID@ key. The problem is that when a validation uses the value of a module combo, and no option is selected in the combo, the validation will use the ID of the default module instead of the one selected in the combo (empty string in this case).

This has been solved by preventing the ID of the default module to be stored in session. Module fields can now be initialized to be initialized with the default module using its column default value.