Fixes issue 23574: Return message of process is shown with grid grouped columns
After processing an order with a column grouped in the grid the record was not being refreshed and the return message was not being showed because there was a javascript exception.

The problem was that if the grid has grouped columns, its data property does not have the updateCacheData function. This has been fixed by only executing that part of the code if the grid.data property has the updateCacheData function.

--HG--
extra : rebase_source : f05f6bbd02328e197ebc3446f9cc8d144c8cee56