Fixes issue 22525, fixes issue 23533. Linked items don't show wrong records
If a sales order was completed, a wrong Return from Customer record appeared in its linked items section. If then a shipment was created with the sales order another wrong record (a return shipment) was shown in its linked items section. This happened because the sql where clause of the headers tabs of the Return Material Receipt and Return from Customer windows were not taking into account that the document type of the record had to be a return.