Related to issue 23870: Foreign key property field values are shown in the grid
The problem was that the datasource was returning the identifier of the property field like this:
- shipmentReceipt.documentType$_identifier: 'MM Shipment'
instead of like this:
- shipmentReceipt$documentType$_identifier: 'MM Shipment'

When a new value is added in the DataToJsonConverter, the dots in the property names are replaced with '$' (DalUtil.FIELDSEPARATOR). There were some places were this replacement was not being done.

--HG--
extra : rebase_source : 4b29d29e424457b0bddd93224a205dcfca6b839a